+++ test/bundlebd.config.js
@@ -1,4 +1,5 @@
 const { defineConfig } = require("..");
+const path = require("path");
 
 module.exports = defineConfig({
 	input: "src",
@@ -8,5 +9,8 @@ module.exports = defineConfig({
 	},
 	format: {
 		indent: "  ",
+	},
+	generateCSSModuleScopedName: (plugin, name, file) => {
+		return plugin + "__" + path.basename(file).split(".")[0] + "--" + name;
 	},
 });